Check if .git directory exist before making the update

+++ scripts/update.sh
@@ -37,6 +37,12 @@ fi
 
 . "$REQUIRE_FILE"
 
+# Check for wallabag .git folder
+if [ ! -d .git ]; then
+    echo "Can not update because wallabag wasn't installed using git (see https://doc.wallabag.org/en/admin/upgrade.html#upgrade-on-a-shared-hosting). Aborting.";
+    exit 2;
+fi
+
 rm -rf var/cache/*
 git fetch origin
 git fetch --tags
